+++ kzg10-rust/kzg/src/lib.rs
@@ -15,6 +15,7 @@ const MCLBN_COMPILED_TIME_VAR: c_int = MCLBN_FR_UNIT_SIZE as c_int * 10 + MCLBN_
 
 #[macro_escape] pub mod init_def;
 pub mod mlc_methods;
+pub mod utilities;
 pub mod data_types {
     pub mod fr;
     pub mod fp;
@@ -24,5 +25,7 @@ pub mod data_types {
     pub mod gt;
 }
 pub mod kzg10;
-pub mod utilities;
+pub mod fk20_fft;
+pub mod zero_poly;
+
 pub mod old;
